requirements/local.txt
-r ./base.txt

black==21.10b0
django-debug-toolbar==3.2.2
django-stubs==1.9.0
flake8==3.9.2
flake8-django==1.1.2
ipdb==0.13.4
ipython==7.29.0
isort==5.10.1
mypy==0.910
pre-commit==2.15.0
pytest-django==4.4.0
